requirements.txt
fastapi 
uvicorn
sqlmodel
psycopg2-binary
pydantic-settings
alembic